Taking input part made clear
